Changes queries to allow NULL.
BLOB fields can't have a default value. This can be handled in the DataWriter no problem. The problem exists when an add-on is disabled (rather than uninstalled). The end result is a column exists in the database that has no default value and there's no code to give it a default value.

The solution here is to actually allow NULL in these fields. Therefore, when new records are added to that table, and there is no data for those fields, they will automatically be NULL with no adverse effects.

At the moment, users will get a "Mysqli statement execute error : Field '3ps_cmfu_options' doesn't have a default value" error if the add-on is disabled.